Move queries to Cassandra hierarchy
