+++ services/galley/src/Galley/Cassandra/Code.hs
@@ -23,7 +23,7 @@ where
 import Brig.Types.Code
 import Cassandra
 import Galley.Cassandra.Store
-import qualified Galley.Data.Queries as Cql
+import qualified Galley.Cassandra.Queries as Cql
 import Galley.Data.Types
 import Galley.Effects.CodeStore (CodeStore (..))
 import Imports
