+++ services/galley/src/Galley/Cassandra/ConversationList.hs
@@ -27,10 +27,10 @@ import Data.Id
 import Data.Qualified
 import Data.Range
 import Galley.Cassandra.Paging
+import qualified Galley.Cassandra.Queries as Cql
 import Galley.Cassandra.ResultSet
 import Galley.Cassandra.Store
 import Galley.Data.Instances ()
-import qualified Galley.Data.Queries as Cql
 import Galley.Effects.ListItems
 import Imports hiding (max)
 import Polysemy
