+++ services/galley/src/Galley/Cassandra/Queries.hs
@@ -15,7 +15,7 @@
 -- You should have received a copy of the GNU Affero General Public License along
 -- with this program. If not, see <https://www.gnu.org/licenses/>.
 
-module Galley.Data.Queries where
+module Galley.Cassandra.Queries where
 
 import Brig.Types.Client.Prekey
 import Brig.Types.Code
