+++ services/galley/src/Galley/Cassandra/SearchVisibility.hs
@@ -19,9 +19,9 @@ module Galley.Cassandra.SearchVisibility (interpretSearchVisibilityStoreToCassan
 
 import Cassandra
 import Data.Id
+import Galley.Cassandra.Queries
 import Galley.Cassandra.Store
 import Galley.Data.Instances ()
-import Galley.Data.Queries
 import Galley.Effects.SearchVisibilityStore (SearchVisibilityStore (..))
 import Galley.Types.Teams.SearchVisibility
 import Imports
